[feat] Add analytics api

+++ event_app/utils.py
@@ -26,7 +26,7 @@ def send_email(emails, subject, message):
 def upload_file(request, file, container):
     if file is None:
         return None
-    file_name = request.User.username + str(uuid.uuid4()) + file.name
+    file_name = request.User.username + str(uuid.uuid4()) + file.name.replace(' ', '')
     blob_client = STORAGE_CLIENT.get_blob_client(container=container, blob=file_name)
     blob_client.upload_blob(file.read())
     return f"https://storageeventmanager.blob.core.windows.net/{container}/{file_name}"
@@ -35,17 +35,21 @@ def upload_file(request, file, container):
 def delete_file(url):
     if url is None:
         return
-    container, file_name = get_container_and_name(url)
-    blob_client = STORAGE_CLIENT.get_blob_client(container=container, blob=file_name)
-    try:
-        blob_client.delete_blob(delete_snapshots="include")
-    except azure_exc.ResourceNotFoundError:
-        pass
+    resolved, container, file_name = get_container_and_name(url)
+    if resolved:
+        blob_client = STORAGE_CLIENT.get_blob_client(container=container, blob=file_name)
+        try:
+            blob_client.delete_blob(delete_snapshots="include")
+        except azure_exc.ResourceNotFoundError:
+            pass
 
 
 def get_container_and_name(url):
     pat = re.compile(r"https://storageeventmanager\.blob\.core\.windows\.net/(.*)/(.*)")
-    return pat.findall(url)[0]
+    try:
+        return True, pat.findall(url)[0]
+    except IndexError:
+        return False, None, None
 
 
 def pro_required(func):
